Add support for HTTP File Upload (cluster-only)
This adds a docker volume to each server, configures each Openfire to use that volume as the backend storage for HTTP File Upload, and configures Openfire to use an endpoint _on the load balancer_ as the endpoint for all HTTP file upload web-interaction.

Note that the httpfileupload plugin is not added in this commit (as its implementation is still being worked on).