chore: Copy java.security from Openfire to configure OCSP
This file is being added to Openfire so I'm porting it into this project as we use a copy of the security directory and overwrite the original.

To pick up the setting in this new file, and the new system property (jdk.tls.server.enableStatusRequestExtension=true), we must also use the updated openfire.sh start script which includes the new file and sets the mentioned system property. This will happen naturally when those changes make it into the Openfire Docker image. Until then you can recreate this behaviour by adding the following to xmpp1 and xmpp2:

```
environment:
    JAVA_TOOL_OPTIONS: >
      -Djava.security.properties=/var/lib/openfire/conf/security/java.security
      -Djdk.tls.server.enableStatusRequestExtension=true
```